moved views into own files
